refactor: relocate brc e2e suite and playwright config into sites/brc-analytics (#1637)
