+++ .github/workflows/run-checks.yml
@@ -21,13 +21,13 @@ jobs:
       - name: Install Playwright browsers
         run: npx playwright install --with-deps
       - name: Run e2e tests
-        run: npm run test:e2e
+        run: npm run test:e2e:brc
       - name: Upload test results
         uses: actions/upload-artifact@v4
         if: failure()
         with:
           name: playwright-report
-          path: playwright-report/
+          path: sites/brc-analytics/playwright-report/
           retention-days: 7
 
   ga2-smoke-tests:
