+++ package.json
@@ -24,7 +24,7 @@
     "lint:python": "./scripts/lint-python.sh",
     "prepare": "husky",
     "test": "jest --runInBand --passWithNoTests",
-    "test:e2e": "npx playwright test",
+    "test:e2e:brc": "npx playwright test --config=sites/brc-analytics/playwright.config.ts",
     "test:e2e:api": "npx playwright test --config=playwright.api.config.ts",
     "test:e2e:ga2": "npx playwright test --config=sites/ga2/playwright.config.ts",
     "select-ga2-organism-images": "python3 -m catalog.ga2.build.py.select_images_for_organism",
